Implements a force-directed formation algorithm for the flock, that doesn't work.
The flyers bunch together around a circle, and when two circles approach each other they get sucked into each other's gravity well.
There's also a bug with the way we've been computing neighbors, we're not taking into account the wrap-around at the edges, so when a bunch of flyers reaches an edge, strange things happen.